pyproject.toml
[tool.poetry]
name = "social-media-post-generator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.17"
langchain-community = "^0.0.36"
youtube-transcript-api = "^0.6.2"
ipykernel = "^6.29.4"
pytube = "^15.0.0"
langchain-google-genai = "^1.0.3"
python-dotenv = "^1.0.1"
streamlit = "^1.35.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
